Add new lib Template method "render" to replace "render_to_file"
- Version bump
- Update README.md and add doc tests
- Add CHANGELOG.md for library
