+++ CONTRIBUTING.md
@@ -5,7 +5,7 @@
 Ensure you have make installed. Run `command -v make`. If there is no
 stdout, search for ways to install `make` on your system.
 
-```shell
+```sh
 make install # installs rust/cargo and cargo crate deps
 ```
 
@@ -13,12 +13,20 @@ make install # installs rust/cargo and cargo crate deps
 
 To generate a release binary to `./target/release/tinted-builder-rust`, run:
 
-```shell
+```sh
 make build
 ```
 
 ## Testing
 
-```shell
-make test
+### CLI
+
+```sh
+make cli_test
+```
+
+### Library
+
+```sh
+make lib_test
 ```
