+++ tinted-builder/CHANGELOG.md
@@ -1,0 +1,9 @@
+# Changelog
+
+## 0.1.0 - 2024-05-03
+
+### Added
+
+- Split tinted-builder-rust library into its own package
+- Add new `Template::render` method
+- Deprecate `Template::render_to_file` method
