--- tinted-builder/src/library/mod.rs
@@ -1,3 +1,0 @@
-pub mod constants;
-pub mod scheme;
-pub mod template;
